+++ packages/Peptides/src/utils/types.ts
@@ -1,5 +1,6 @@
 import * as DG from 'datagrok-api/dg';
 import {SCALING_METHODS} from './constants';
+import {ClusterType} from '../model';
 
 export type DataFrameDict = {[key: string]: DG.DataFrame};
 
@@ -11,6 +12,7 @@ export type PositionToAARList = {[postiton: string]: string[]};
 
 export type MonomerSelectionStats = {[position: string]: {[monomer: string]: number}};
 export type MonomerPositionPair = {monomer: string, position: string};
+export type Cluster = {name: string, type: ClusterType};
 
 export type PeptidesSettings = {
   sequenceColumnName?: string,